Ensures fs.stat and fs.type are asynchronous
