+++ lute/fs/src/fs_impl.cpp
@@ -1,5 +1,6 @@
 #include "fs_impl.h"
 
+#include "lute/time.h"
 #include "lute/UVRequest.h"
 
 #include "lua.h"
@@ -258,6 +259,145 @@ int remove_impl(lua_State* L, const char* path)
                 [](lua_State* L)
                 {
                     return 0;
+                }
+            );
+        }
+    );
+
+    return lua_yield(L, 0);
+}
+
+static const char* fileModeToType(uint64_t mode)
+{
+    if (S_ISDIR(mode))
+    {
+        return UV_TYPENAME_DIR;
+    }
+    else if (S_ISREG(mode))
+    {
+        return UV_TYPENAME_FILE;
+    }
+    else if (S_ISCHR(mode))
+    {
+        return UV_TYPENAME_CHAR;
+    }
+    else if (S_ISLNK(mode))
+    {
+        return UV_TYPENAME_LINK;
+    }
+#ifdef S_ISBLK
+    else if (S_ISBLK(mode))
+    {
+        return UV_TYPENAME_BLOCK;
+    }
+#endif
+#ifdef S_ISFIFO
+    else if (S_ISFIFO(mode))
+    {
+        return UV_TYPENAME_FIFO;
+    }
+#endif
+#ifdef S_ISSOCK
+    else if (S_ISSOCK(mode))
+    {
+        return UV_TYPENAME_SOCKET;
+    }
+#endif
+    else
+    {
+        return UV_TYPENAME_UNKNOWN;
+    }
+}
+
+static int createDurationFromTimespec32(lua_State* L, uv_timespec_t timespec)
+{
+    uv_timespec64_t extended{static_cast<int64_t>(timespec.tv_sec), static_cast<int32_t>(timespec.tv_nsec)};
+    return createDurationFromTimespec(L, extended);
+}
+
+int stat_impl(lua_State* L, const char* path)
+{
+    uvutils::ScopedUVRequest<FSRequest> req(L);
+    uv_fs_stat(
+        uv_default_loop(),
+        &req->req,
+        path,
+        [](uv_fs_t* req)
+        {
+            auto r = uvutils::retake<FSRequest>(req);
+            auto result = req->result;
+
+            if (result < 0)
+            {
+                r->fail("Error getting metadata of file %s: %s", req->path, uv_strerror(result));
+                return;
+            }
+
+            r->succeed(
+                [stat = req->statbuf](lua_State* L)
+                {
+                    lua_createtable(L, 0, 6);
+
+                    auto type = fileModeToType(stat.st_mode);
+                    lua_pushstring(L, type);
+                    lua_setfield(L, -2, "type");
+
+                    // this is fine unless the file is 9 petabytes
+                    lua_pushnumber(L, static_cast<double>(stat.st_size));
+                    lua_setfield(L, -2, "size");
+
+                    createDurationFromTimespec32(L, stat.st_birthtim);
+                    lua_setfield(L, -2, "created");
+
+                    createDurationFromTimespec32(L, stat.st_atim);
+                    lua_setfield(L, -2, "accessed");
+
+                    createDurationFromTimespec32(L, stat.st_mtim);
+                    lua_setfield(L, -2, "modified");
+
+                    // permissions
+                    lua_createtable(L, 0, 2);
+
+                    // libuv writes this correctly cross-platform
+                    bool canAnyWrite = stat.st_mode & 0222;
+                    lua_pushboolean(L, !canAnyWrite);
+                    lua_setfield(L, -2, "readonly");
+
+                    lua_setfield(L, -2, "permissions");
+
+                    return 1;
+                }
+            );
+        }
+    );
+
+    return lua_yield(L, 0);
+}
+
+int type_impl(lua_State* L, const char* path)
+{
+    uvutils::ScopedUVRequest<FSRequest> req(L);
+    uv_fs_stat(
+        uv_default_loop(),
+        &req->req,
+        path,
+        [](uv_fs_t* req)
+        {
+            auto r = uvutils::retake<FSRequest>(req);
+            auto result = req->result;
+
+            if (result < 0)
+            {
+                r->fail("Error getting type of file %s: %s", req->path, uv_strerror(result));
+                return;
+            }
+
+            r->succeed(
+                [stat = req->statbuf](lua_State* L)
+                {
+                    auto type = fileModeToType(stat.st_mode);
+                    lua_pushstring(L, type);
+                    return 1;
                 }
             );
         }
